renamed s3 lib file
