+++ omniscripts/benchmarks/census/census_pandas_modin.py
@@ -4,7 +4,7 @@ from sklearn import config_context
 import warnings
 from timeit import default_timer as timer
 
-from omniscripts.tools.s3load import download_folder
+from omniscripts.tools.s3_load import download_folder
 from omniscripts import BaseBenchmark, BenchmarkResults
 from omniscripts.pandas_backend import pd
 from omniscripts.benchmark_utils import (
